refactor(Player): create Interface Player
