+++ src/main/kotlin/blackjack/domain/Bat.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.domain.player.Dealer
 import blackjack.domain.state.Blackjack
 import blackjack.domain.state.Bust
 import blackjack.domain.state.State
