+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -1,14 +1,12 @@
-package blackjack.domain
+package blackjack.domain.player
 
 import blackjack.domain.card.Card
 import blackjack.domain.state.Running
 import blackjack.domain.state.State
 
-open class Player(val name: String, state: State) {
+abstract class Player(val name: String, state: State) {
     open var state: State = state
         protected set
-
-    lateinit var bat: Bat
 
     open fun draw(card: Card) {
         state = state.draw(card)
@@ -18,11 +16,6 @@ open class Player(val name: String, state: State) {
         state = state.stay()
     }
 
-    fun bat(money: Double) {
-        bat = Bat(money)
-    }
-
     fun canDraw() = state is Running
     fun score() = state.score()
-    fun profit(dealer: Dealer) = bat.profit(state, dealer)
 }
