+++ src/test/kotlin/blackjack/domain/player/DealerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Denomination
