+++ src/test/kotlin/blackjack/domain/player/UserTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Denomination
@@ -8,12 +8,12 @@ import io.kotest.matchers.collections.shouldContainInOrder
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 
-internal class PlayerTest {
+internal class UserTest {
     @Test
     fun `플레이어는 이름과 핸드를 가진다`() {
         val firstCard = Card(Shape.CLUB, Denomination.ACE)
         val secondCard = Card(Shape.DIAMOND, Denomination.FOUR)
-        val player = Player("최원준", FirstTurn.draw(firstCard, secondCard))
+        val player = User("최원준", FirstTurn.draw(firstCard, secondCard))
 
         player.name shouldBe "최원준"
         player.state.cards() shouldContainInOrder listOf(firstCard, secondCard)
@@ -23,7 +23,7 @@ internal class PlayerTest {
     fun `플레이어는 카드 합을 가진다`() {
         val firstCard = Card(Shape.CLUB, Denomination.TEN)
         val secondCard = Card(Shape.DIAMOND, Denomination.FOUR)
-        val player = Player("최원준", FirstTurn.draw(firstCard, secondCard))
+        val player = User("최원준", FirstTurn.draw(firstCard, secondCard))
 
         player.score() shouldBe 14
     }
